say bye to coordinator under 18
